Decrypt password file with a key
It turns out an encrypted string needs to be encrypted with a custom key, otherwise PShell uses the user's account and the password file can't be used on a CI machine. Adding a password file encrypted with a 16-byte key for research.

We'll want a secure password file for production.