Fixed error in resource group creation
When creating a resource group, we embed a hashtable that had key values meant to uniquely identify the resource group. This was to work around limitations with the Find-AzureRmResourceGroup cmdlet. However, I forgot to remove hard coded values so simultaneous deployments failed as agents thought the desired resource group already existed.